[EFM32] Fix integer division bug in pwm read.

+++ libraries/mbed/targets/hal/TARGET_Silicon_Labs/TARGET_EFM32/TARGET_EFM32GG_STK3700/pwmout_api.c
@@ -95,7 +95,7 @@ void pwmout_write(pwmout_t* obj, float value) {
 }
 
 float pwmout_read(pwmout_t* obj) {
-    return obj->width_cycles / obj->period_cycles;
+    return obj->width_cycles / (float) obj->period_cycles;
 }
 
 // Set the PWM period, keeping the absolute pulse width the same.
